refactor(cmd): split main.go into cohesive composition-root files
Break the 1290-line cmd/ovumcy/main.go into sibling files in package main,
each a cohesive unit: config.go (runtime config types + env resolution),
server.go (fiber app, middleware, static, http error, security headers),
ratelimit.go (proxy-trust matcher, key generator, limit handlers), env.go
(env parsing primitives), version.go (build identity + asset cache-bust),
commands.go (operator CLI dispatch), startup.go (startup logging).

main.go keeps only the composition root and process lifecycle: main(),
runServer, startReminderScheduler, closeDatabase, the must* constructors, and
the graceful-stop machinery (installGracefulShutdown/retryShutdown stay in
main.go, per the deployment guide's boot-window-race anchor).

Pure declaration move within one package: 54 funcs / 7 types preserved, no
logic change. go build/vet/test ./cmd/ovumcy green, gofmt clean.

Refs #251
